requirements.txt
accelerate==0.34.2
altair==5.3.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
filelock==3.15.4
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.24.5
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.3
psutil==6.0.0
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
safetensors==0.4.5
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
streamlit==1.37.1
sympy==1.13.3
tenacity==8.5.0
tokenizers==0.20.0
toml==0.10.2
toolz==0.12.1
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
